requirements.txt
selenium==3.12.0
Pillow==5.1.0
